Add PINNIPED_USE_LOCAL_KIND_REGISTRY env var
- ensures regular use of hack/prepare-for-integration-tests.sh
  - PINNIPED_USE_LOCAL_KIND_REGISTRY=1  ./hack/prepare-for-integration-tests.sh --clean --alternate-deploy ./hack/noop.sh --post-install ./hack/build-carvel-packages.sh
  - ./hack/prepare-for-integration-tests.sh --clean

+++ hack/kind-down.sh
@@ -8,9 +8,11 @@ set -euo pipefail
 ROOT="$( cd "$( dirname "${BASH_SOURCE[0]}" )/.." && pwd )"
 cd "${ROOT}"
 
-reg_name='kind-registry.local'
-docker network disconnect "kind" "${reg_name}" || true
-docker stop "${reg_name}" || true
-docker rm "${reg_name}" || true
+if [[ "${PINNIPED_USE_LOCAL_KIND_REGISTRY:-}" != "" ]]; then
+  reg_name='kind-registry.local'
+  docker network disconnect "kind" "${reg_name}" || true
+  docker stop "${reg_name}" || true
+  docker rm "${reg_name}" || true
+fi
 
 kind delete cluster --name pinniped
